@@ -1,12 +1,16 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import './css/index.min.css'
 import 'devicon/devicon.min.css'
 import App from './App'
 
-ReactDOM.render(
+const container = document.getElementById('root')
+if (!container) {
+  throw new Error("The element #root wasn't found")
+}
+const root = createRoot(container)
+root.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 )
